server: fold self-bootstrap into init-bootstrap code
The two methods for bootstrapping a new cluster are starting an
uninitialized node without join flags and starting an uninitialized node
with joins flags followed by a BootstrapCluster RPC to it. The former
can easily be modeled by the latter, so let's do that: if we're started
without join flags, start the init server and send an RPC to it.

This unifies our bootstrap code and makes it simpler to think about.

Release note: None