roachprod: fail to expand when not all nodes are known
This PR adds error propagation to the roachprod `expander` and ensures that
we return an error when the node spec being used for expansion does not
match the shape of the cluster.

Why the data in the SyncedCluster doesn't match reality remains a mystery but
this should at least lead to better error messages.

Release note: None