util/quotapool,storage/proposalquota: generalize the storage.quotaPool
This commit takes that existed storage/quota_pool.go and generalized it to
enable more implementations of quota pools with different logic for how
acquistion decisions are made. It then implements the interface as currently
used by the storage package to control the rate of outgoing proposals.

The bulk of the complexity comes from the usual challenge when abstracting
data structures in go, reducing allocations due to boxing in interfaces.

Release note: None