metric: add Summary metric which tracks a trailing distribution of values
A summary tracks a trailing distribution of values. It is similar in principle
to a histogram with a different set of tradeoffs. Prometheus has support for
summaries but generally recommends against them in the common case for reasons
outlined below.

Pros:
  - No configuration for the value range.
  - Generally much higher precision, often 10s of PPM rather than 2 sig-figs
    over a very wide range.
  - Amenable to expotential trailing summaries.
    - Histograms might be too, the way we do windowing now is a bit
      underconsidered.
  - Data structure is optimized concurrent modification.
    - More of an implementation detail than anything, could imagine optimizing
      HDRHistogram similarly.
  - Smaller serialized footprint than HDRHistogram for much higher precision.
    - The upper bound of the serialized size is controlled by the Compression
      factor which today defaults to 128. The size of a serialized TDigest is
      less than ~16*Compression so by default a TDigest will be less than 2kb.

Cons:
  - Larger in-memory footprint than HDRHistogram.
    - The TDigest data structure uses memory o(16*Compression*(1+BufferFactor))
      bytes which means at the default setting of 12kb whereas an HDRHistogram
      like we use for latency defaults to 4kb but we keep two of them (windowed
      and cumulative) so the total in memory footprint of a histogram metric is
      ~8kb making this not much bigger.
  - No out-of-the-box merging of data between nodes by prometheus.
    - See [the Prometheus docs on Histograms vs Summaries](https://prometheus.io/docs/practices/histograms/).
    - These things are mergeable and require less data to be shipped around for
      more precision but there's currently no out of the box support in either
      our own tsdb or in prometheus (I intend to look at adding support to
      prometheus at some point).

Another potential point of contention with this PR is that it adds a dependency
on a personal repository github.com/ajwerner/tdigest. I would be happy to move
this over and/or get it reviewed.

The main motivation for adding this metric type is to drive request size
approximation for use in a read quota. The TDigest is much cheaper to use in a
best-effort way to pick a value at a given quantile than is a histogram which
today requires locking the data structure and allocating the `Bars`. I imagine
that the HDRHistogram could be optimized to better provide this functionality
but given I've already written the TDigest, may as well use it for situations
where it fits pretty well.

Using a summary like this may also be nice for the SQL latency we display which
carries incredibly low precision at the top of the distribution where this data
structure would have high precision at the top of the distribution.

Release note: None