distsqlrun: always attach a local recording span to processors
This commit ensures that when a processor is started it always has a recordable
span in its context. Armed with this assumption, always enable table reader
stats recording into the context span.

My understanding is that given stats in the context span and the finishFunc
we'll always end up sending the TableReaderStats inside the
ProducerMeta.TraceData.

Release note: None