ui: upgrade jsdoc version to allow Node 12
Generating TypeScript protobuf definitions was failing on Node 12,
because pbts was depending on an old version of jsdoc, which was in turn
depending on a library called requizzle, which was incompatible with
Node 12. Bump the version of jsdoc in the lockfile so we can
transitively pick up a newer version of requizzle that does support
Node 12.

See also: protobufjs/protobuf.js#1212

Release note: None