opt: calculate cardinality from filters
This commit adds logic to determine whether a tight cardinality bound
can be determined from the filters in a Select or Scan expression, and
updates the cardinality accordingly in the logical properties.
Specifically, it may be possible to determine a tight bound if the key
column(s) are constrained to a finite number of values.

Release note: None