roachtest: tune to fix flake
We've been observing relatively regular OOMs at the current concurrency.

This commit fixes an egregious bug where the warehouse count was specifying the
concurrency. It then increase the concurrency for AWS and decreases it for GCP.

I have verified that this still fails on 19.1 and passes on

I've verified that this still fails on v19.1.4.

Closes #39684.

Release note: None