storage: deflake TestSnapshotAfterTruncationWithUncommittedTail
Fixes #40608. Between sending out the increment and log truncation
request, the leaseholder may have changed from underneath us. Since the
range in question is using an expiration based lease, when we run into a
NotLeaseHolderError and don't correct for it, we're stuck trying to send
the request to a non-leaseholder.

Release note: None