storage/gc: carve out a storage subpackage for gc
This commit moves the logic of RunGC as well as the engine.GC struct into a
separate subpackage. As of this commit that subpackage contains no testing
other than what existed for the engine.GC code.

The RunGC function is a reasonably well specified interface to separate
the logic of scanning a range and collecting the garbage from the gcQueue.

I was getting overwhelmed by testing boundaries and unit testing in general.

Release note: None