This made a bit of a mess.
This commit creates a new subpackage of changefeedccl called tablefeed.
A table feed tracks events on tables. The client has the option to filter
out certain table events which it is not interested in.

A tablefeed provides a queue to which the caller can pass a timestamp.
The queue promises to return all events which occurred at or before
the caller's timestamp. If that timestamp has not yet been proven,
the call will block until it has been.

This package will be used by the kvfeed package to perform logical backfill
at the appropriate moment and to ensure that higher levels will be able to
properly interpret any kv-pair sent up to it.

The introduction of this abstraction provides points to improve efficiency.

Ideally we'd only ever have a single goroutine for all changefeeds fetching
table descriptors.

Furthermore we're going to use this layer to introduce locking. We really
don't want to have more than one lock per node on a table.

Release note (<category, see below>): <what> <show> <why>