changefeedccl: rename the poller to kvfeed, carve out subpackages
This commit moves the `*poller` abstraction to a new package called
kvfeed. The poller has not actually performed polling outside of the
table descriptors in several releases. The entrypoint to the
functionality it used to expose is now `kvfeed.Run`. In addition
to the exported method, this refactor introduces several internal
abstractions to aid in testing and readability.

This change also introduces a changefeedbase package to export options
used in subpackages.

This change is large but wholly cosmetic.

Release note: None