sqlbase,catalogkv,*: have GetDescriptorByID return an interface
This commit has GetDescriptorByID return an "unwrapped" descriptor and removes
`UnwrapDescriptor` function from sqlbase. This terminology is all crap given
the `UnwrapDescriptor` function returned a descriptor that is itself sort of
wrapped. Good news, this terminology is going away. There is a hodge-podge
of other unfortunate touch up in this commit.

Release note: None