sqlbase,*: deprecate ID and Name on DatabaseDescriptor
This commit adopts the interface methods for accessing the ID and Name of
a DatabaseDescriptor. The reworking in this commit of previous changes also
in this PR is annoying, I'm sorry to the reviewers. I'm also increasingly
sensing the urgency of eschewing the concrete descriptor wrappers in favor
of interfaces but I'm not going to try to deal with that in this PR.

It is in many ways more of a mess than what was there but it's laying the
foundations.

Release note: None