kv/kvclient/kvcoord: add mechanism to manually set the anchor key
In order for a transaction to be pushed, it must have an anchor key. Normally
transactions are only pushed if another transaction is blocked on a lock owned
by that transaction. However, there are cases where, within other subsystems
of cockroach, transactions are logically blocked on other transactions. This
generally happens in the context of singleflights. The work in this PR is to
inform KV of these logical dependencies so that its deadlock detection can
resolve these scenarios.

This commit introduces a new method to the TxnSender interface: `SetAnchor`.
The new method is intended to be used to allow a transaction to be pushed
even if it has not created any locks.

Epic: none

Release note: None