kv/kvclient/kvcoord: add BlockOn method to TxnSender
This commit adds a mechanism for logical dependencies to be communicated to
kvclient for the purpose of deadlock detection. This is useful in cases where
a transaction is not able to make progress until a separate transaction
finishes. These situations would result in deadlocks if the latter transactions
ends up blocked in a cycle involving the former transaction.

Epic: none

Release note: None